restaurant_order_backend/requirements.txt
asgiref==3.8.1
dataclasses==0.6
dj-database-url==2.3.0
Django==5.2
django-admin-interface==0.30.0
django-admin-rangefilter==0.13.2
django-colorfield==0.13.0
django-cors-headers==4.7.0
django-suit==0.2.28
djangorestframework==3.16.0
djongo==1.2.31
dnspython==2.7.0
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pymongo==4.12.0
python-decouple==3.8
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.13.2
tzdata==2025.2
whitenoise==6.9.0
